Made every layer child have ID properties, not just MapSprites.

+++ core/src/com/bamboo/bridgebuilder/data/GroupMapPolygonData.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 
 public class GroupMapPolygonData extends MapPolygonData
 {
-    public ArrayList<Integer> mapSpriteIDs;
+    public ArrayList<Long> mapSpriteIDs;
 
     public GroupMapPolygonData(){}
     public GroupMapPolygonData(MapPolygon mapPolygon, float offsetX, float offsetY)
